Recopilación de data/Recopilación de data  - Reddit/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
idna==2.10
numpy==1.20.0
pandas==1.2.1
praw==7.1.0
prawcore==1.5.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
update-checker==0.18.0
urllib3==1.26.3
websocket-client==0.57.0
